Add user profile view and leaderboard click handler.
